Enable configuration of logging verbosity through simplelog
The alpaca crate which we use for as the backend for the actual
interaction with the Alpaca API provides various log messages that may
be helpful when debugging issues.
To make these accessible without recompilation of the program, this
change introduces a new option, -v/--verbose, which increments the log
level every time it is provided.