Add support for listing positions
This change adds support for listing currently held positions. To
accomplish that we introduce a new subcommand, position, and within it
another one, list. That is done mostly for symmetry with the structure
of the order subcommand; right now there are no plans to add additional
functionality to the position subcommand.
Formatting the positions is not quite trivial as we have to ensure
proper alignment of all values in order to make the output appealing and
concise at the same time.