Take care of exiting with appropriate code manually
So far we have relied on Rust's std::process::Termination trait for
reporting an appropriate exit code to callers, depending on the result
of the performed operation.
This trait has always been a pain to work with. It starts with the
policy it enforces of unconditionally prefixing errors with an "Error: "
followed by a quoted version of the error. It continues with the trait
being not yet available to be implemented for custom types on stable
Rust.
Hence, with this change we take care of manually invoking exit with the
appropriate exit code, depending on the result of the program, instead
of just returning the result from main.