Rework error reporting once more
This change reworks our error reporting some more. Instead of bubbling
up an error until it is eventually handled in the program's main
function, this change causes errors to be printed immediately. This step
has the advantage that we can easily print multiple error messages (in
case multiple requests failed, for example), as it is a real pain to
merge multiple errors using future combinators.
Because we are nothing but a command line application that can't do much
more than print to stdio channels anyway, we may as well do that in more
than one location.