Print information about total value & gains in position listing
The position listing, while providing a very nice and concise overview
of the open positions, is missing important information relevant to most
of not all users: the percentage gains for the day and in total.
This change takes care of printing this information as part of the
listing. Included are also absolute values, i.e., the total amount of
funds invested and what the equity purchased is worth right now.