Add support for submitting orders based on a certain value
This change adds support for submitting orders based on a certain amount
of funds to invest. In order to submit such an order we first retrieve
the current market value of the asset to purchase and then divide the
provided value by that price, rounding down the final result. We use
Alpaca's market data API which is known to have fairly unreliable data,
but the hope is that it will be good enough, as the requirements here
are very low.