Add support for listing orders based on status
So far we have only support listing open orders, because that's the
default that our API bindings were using -- without any means to change.
However, we have come to the point where it would be great to also list
closed orders.
To facilitate such use cases, we have adjusted the alpaca crate to
include a status member in the orders request object. This change pulls
in this new version of the crate and adds a new option, --status, to the
orders list command.