Change --status option to order list command to --closed
When initially devised, we tried to expose all available functionality
when it comes to listing orders through the CLI. That means we were able
to list open orders, closed ones, and the combination of both. The
problem with that is that because of the way we format orders, it is not
obvious what status an order has. Now, we could certainly change that,
but ultimately the use-case we really wanted to enable was to list
closed orders as well.
To that end, this change replaces the --status option to the order list
command with a simpler --closed, which will cause only closed orders to
be listed (as opposed to open ones, which is the default).