Add support for generating a bash completion script
This change adds support for generating a bash completion script. If
sourced, the shell will provide tab completions for the program's
arguments.
There are two possible approaches provided by clap for going about
generating shell completion functionality: either at build time by
separately generating the clap parsers out-of-band or at run time, as an
option to the main program itself. We are generally not too much in
favor of a run time approach, as it means less inspectability at
installation time and more overhead in the form of code crammed into the
main binary.
Hence, with this change we take the "build time" approach. Clap
recommends hooking the generation up in build.rs, but this seems like an
inflexible choice. For one, that is because it would mean
unconditionally generating this file or some user unfriendly environment
variable based approach to making the process conditional, but also
because specifying the command for which to generate the script should
likely be configurable. That is a limitation of the completion script
that clap generates (see https://github.com/clap-rs/clap/issues/1764).
Instead, we provide a utility program that emits the completion script
to standard output, accepting regular command line options itself.